% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_metric_calc.R
\name{add_latest_dummy_point}
\alias{add_latest_dummy_point}
\title{Internal helper function for adding a last dummy point (for each group) which marks the end of the last
interval (of each group).}
\usage{
add_latest_dummy_point(dataset, grouping_var, time_type, time_unit)
}
\arguments{
\item{dataset}{\verb{[data.frame]} Data frame returned by fill_missing_intervals() function.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the "dm_dataset". The
values of this column define the different groups in the plot.}

\item{time_type}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based). Either "date" or "day".}

\item{time_unit}{\verb{[character(1)]} Determining the binsize. Either "week" or "month".}
}
\value{
A data frame which contains an additional dummy row for each group after the last real interval.
}
\description{
This is necessary for drawing the last interval. Otherwise there would be just a point for the last interval, because
an interval is drawn from the starting point to the x coordinates of the starting point of the next interval and
if there would be no next starting point, the interval is not drawn.
}
\keyword{internal}
