% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{create_plot}
\alias{create_plot}
\title{Internal helper function for creating the ggplot2 base for the dynamic and interactive plot.}
\usage{
create_plot(
  dm,
  ds,
  ae,
  subjid_var,
  grouping_var,
  selected_levels,
  metric_selection,
  time_selection,
  unit_selection,
  tooltip_decimal_places,
  x_step_size_list,
  color_palette,
  show_ungrouped
)
}
\arguments{
\item{dm}{\verb{[data.frame]} The demographics data frame.}

\item{ds}{\verb{[data.frame]} The disposition data frame returned by the prepare_ds_data() function.}

\item{ae}{\verb{[data.frame]} The adverse event data frame returned by the prepare_ae_data() function.}

\item{subjid_var}{\verb{[character(1)]} Character name of the unique subject identifier column in all datasets.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the dm data frame. The
values of this column define the different groups in the plot.}

\item{selected_levels}{\verb{[character(0+)]} Character vector containing the currently selected grouping levels that
should be displayed in the plot.}

\item{metric_selection}{\verb{[character(1)]} For determining the calculated rate on the y-axis. Either
METRIC_BUTTONS_OPT1 or METRIC_BUTTONS_OPT2.}

\item{time_selection}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based). Either TIMETYPE_BUTTONS_OPT1 or TIMETYPE_BUTTONS_OPT2.}

\item{unit_selection}{\verb{[character(1)]} Determining the binsize when computing the reporting rate per active
patients. Either BINSIZE_BUTTONS_OPT1 or BINSIZE_BUTTONS_OPT2.}

\item{tooltip_decimal_places}{\verb{[integer(1)]} Number of decimal places to round the value shown in the hovertext.}

\item{x_step_size_list}{\verb{[list]} Named list specifying the tick interval (step size) on the x-Axis depending on the
time unit. Composed of:
\itemize{
\item days \verb{[integer(1)]}
Integer value specifying the tick interval in days when the x-axis is displayed in days.
\item weeks \verb{[integer(1)]}
Integer value specifying the tick interval in weeks when the x-axis is displayed in weeks.
\item months \verb{[integer(1)]}
Integer value specifying the tick interval in months when the x-axis is displayed in months.
}}

\item{color_palette}{\verb{[named character(1+)]} Named character vector of color values, where the names correspond to
the levels of the grouping variable. Always contains the entry "ungrouped" which is fixed to the value "#000000".}

\item{show_ungrouped}{\verb{[logical]} Specifying if the ungrouped data should be displayed when a grouping variable is
selected.}
}
\value{
A \verb{[ggplot2::ggplot]} object. A static ggplot-based visualization showing calculated rates over time
based on the selected grouping variable and levels, metric, and time unit.
}
\description{
Internal helper function for creating the ggplot2 base for the dynamic and interactive plot.
}
\keyword{internal}
