% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_columns_with_labels}
\alias{get_columns_with_labels}
\title{Internal helper function for getting column names with their labels for the UI grouping dropdown.}
\usage{
get_columns_with_labels(dataset, columns)
}
\arguments{
\item{dataset}{\verb{[data.frame]} The demographics dataframe which contains the possible variables for grouping and their
labels.}

\item{columns}{\verb{[character(0+)]} Character vector of the columnnames that should be returned with their label.}
}
\value{
Named character vector where values are the original columnnames and the names are the string to display in
the format "columnname (columnlabel)" or only "columnname" if no label exsits.
}
\description{
Internal helper function for getting column names with their labels for the UI grouping dropdown.
}
\keyword{internal}
