% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_geom_point}
\alias{get_geom_point}
\title{Internal helper function for getting the right geom_point_interactive depending on the grouping setting.}
\usage{
get_geom_point(dataset, grouping_var = NULL)
}
\arguments{
\item{dataset}{\verb{[data.frame]} A data frame returned by prepare_plot_data() function.}

\item{grouping_var}{\verb{[character(1) | NULL]} Character name of the grouping variable (column) in the dm data frame.
The values of this column define the different groups in the plot.}
}
\value{
A ggiraph::geom_point_interactive() layer object configured for either grouped or ungrouped data depending
on grouping_var.
}
\description{
Internal helper function for getting the right geom_point_interactive depending on the grouping setting.
}
\keyword{internal}
