% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hovertext.R
\name{get_hovertext}
\alias{get_hovertext}
\title{Internal helper function to create the dynamic hovertext that is displayed when hovering over the interactive plot.}
\usage{
get_hovertext(
  dataset,
  grouping_var,
  metric_selection,
  time_selection,
  unit_selection,
  tt_dec_places
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} Data frame returned by the get_time_for_plot() function.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the demographics data frame.
The values of this column define the different groups in the plot.}

\item{metric_selection}{\verb{[character(1)]} For determining the calculated rate on the y-axis. Either
METRIC_BUTTONS_OPT1 or METRIC_BUTTONS_OPT2.}

\item{time_selection}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based) Either TIMETYPE_BUTTONS_OPT1 or TIMETYPE_BUTTONS_OPT2.}

\item{unit_selection}{\verb{[character(1)]} Determining the binsize when computing the reporting rate per active
patients. Either BINSIZE_BUTTONS_OPT1 or BINSIZE_BUTTONS_OPT2.}

\item{tt_dec_places}{\verb{[integer(1)]} Number of decimal places to round the value shown in the hovertext.}
}
\value{
A processed data frame which now contains an additional column "hovertext".
}
\description{
Internal helper function to create the dynamic hovertext that is displayed when hovering over the interactive plot.
}
\keyword{internal}
