% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hovertext.R
\name{get_hovertext_ae_info}
\alias{get_hovertext_ae_info}
\title{Internal helper function for getting the part of the hovertext which shows the ratio and the number of AEs.}
\usage{
get_hovertext_ae_info(dataset, tt_dec_places)
}
\arguments{
\item{dataset}{\verb{[data.frame]} Data frame returned by the get_time_for_plot() function.}

\item{tt_dec_places}{\verb{[integer(1)]} Number of decimal places to round the value shown in the hovertext.}
}
\value{
A processed data frame with the additional column "ht_ae_info". The values in this column are later a part
of the generated hovertext.
}
\description{
Internal helper function for getting the part of the hovertext which shows the ratio and the number of AEs.
}
\keyword{internal}
