% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hovertext.R
\name{get_hovertext_levels_display}
\alias{get_hovertext_levels_display}
\title{Internal helper function for getting the part of the hovertext that contains the levels information.}
\usage{
get_hovertext_levels_display(dataset, grouping_var)
}
\arguments{
\item{dataset}{\verb{[data.frame]} Data frame returned by the get_time_for_plot() function.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the demographics data frame.
The values of this column define the different groups in the plot.}
}
\value{
A processed data frame with the additional column "ht_levels_display". The values in this column are later
a part of the generated hovertext.
}
\description{
Internal helper function for getting the part of the hovertext that contains the levels information.
}
\keyword{internal}
