% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_indices_for_tickvals}
\alias{get_indices_for_tickvals}
\title{Internal helper function creating a logical vector for the indexing of the ticks on the x-Axis of the plot. Applying
the logic on the tick_text (and not tick_vals) because that's what displayed in the plot in the end.}
\usage{
get_indices_for_tickvals(step_size, tick_text)
}
\arguments{
\item{step_size}{\verb{[integer(1)]} Integer value specifying the tick interval on the x axis.}

\item{tick_text}{\verb{[character(1+)]} A character vector setting the text displayed at the ticks positions.}
}
\value{
A logical vector specifying the indices at which ticks should be shown in the plot on the x axis.
}
\description{
Internal helper function creating a logical vector for the indexing of the ticks on the x-Axis of the plot. Applying
the logic on the tick_text (and not tick_vals) because that's what displayed in the plot in the end.
}
\keyword{internal}
