% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_legend_text}
\alias{get_legend_text}
\title{Internal helper function for getting a dynamic description for the levels in the legend of the plot. Returns the
label of the currently selected grouping variable if there is one. Else it just returns the grouping variable string.}
\usage{
get_legend_text(dataset, grouping_var)
}
\arguments{
\item{dataset}{\verb{[data.frame]} The demographics dataframe which contains the possible variables for grouping and their
labels.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the dm data frame}
}
\value{
A string which is either the label of the grouping variable or the grouping variable itself.
}
\description{
Internal helper function for getting a dynamic description for the levels in the legend of the plot. Returns the
label of the currently selected grouping variable if there is one. Else it just returns the grouping variable string.
}
\keyword{internal}
