% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_plot_datasets}
\alias{get_plot_datasets}
\title{Internal helper function for preparing grouped and/or ungrouped datasets for plotting based on the user selections.
Besides that the function also takes care of the selection of the dataset that should be used for generating the x
Axis. If the ungrouped dataset is available, it is used for the x Axis, because it contains the entire time span.}
\usage{
get_plot_datasets(
  dm,
  ae,
  ds,
  subjid_var,
  grouping_var,
  time_selection,
  metric_selection,
  unit_selection,
  tooltip_decimal_places,
  show_ungrouped
)
}
\arguments{
\item{dm}{\verb{[data.frame]} The demographics data frame.}

\item{ae}{\verb{[data.frame]} The adverse event data frame returned by the prepare_ae_data() function.}

\item{ds}{\verb{[data.frame]} The disposition data frame returned by the prepare_ds_data() function.}

\item{subjid_var}{\verb{[character(1)]} Character name of the unique subject identifier column in all datasets.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the dm data frame. The
values of this column define the different groups in the plot.}

\item{time_selection}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based). Either TIMETYPE_BUTTONS_OPT1 or TIMETYPE_BUTTONS_OPT2.}

\item{metric_selection}{\verb{[character(1)]} For determining the calculated rate on the y-axis. Either
METRIC_BUTTONS_OPT1 or METRIC_BUTTONS_OPT2.}

\item{unit_selection}{\verb{[character(1)]} Determining the binsize when computing the reporting rate per active
patients. Either BINSIZE_BUTTONS_OPT1 or BINSIZE_BUTTONS_OPT2.}

\item{tooltip_decimal_places}{\verb{[integer(1)]} Number of decimal places to round the value shown in the hovertext.}

\item{show_ungrouped}{\verb{[logical]} Specifying if the ungrouped data should be displayed when a grouping variable is
selected.}
}
\value{
A list containing the ungrouped dataset or NULL; the grouped dataset or NULL and the dataset that should be
used for axis calculation.
}
\description{
Internal helper function for preparing grouped and/or ungrouped datasets for plotting based on the user selections.
Besides that the function also takes care of the selection of the dataset that should be used for generating the x
Axis. If the ungrouped dataset is available, it is used for the x Axis, because it contains the entire time span.
}
\keyword{internal}
