% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_x_axis_label}
\alias{get_x_axis_label}
\title{Internal helper function for getting the desired label of the x-axis, depending on the inputted timetype and metric.}
\usage{
get_x_axis_label(metric_selection, time_selection, unit_selection)
}
\arguments{
\item{metric_selection}{\verb{[character(1)]} For determining the calculated rate on the y-axis. Either
METRIC_BUTTONS_OPT1 or METRIC_BUTTONS_OPT2.}

\item{time_selection}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based). Either TIMETYPE_BUTTONS_OPT1 or TIMETYPE_BUTTONS_OPT2.}

\item{unit_selection}{\verb{[character(1)]} Determining the binsize when computing the reporting rate per active
patients. Either BINSIZE_BUTTONS_OPT1 or BINSIZE_BUTTONS_OPT2.}
}
\value{
The label for the x Axis. Either "Date", "Days", "Weeks" or "Months".
}
\description{
Internal helper function for getting the desired label of the x-axis, depending on the inputted timetype and metric.
}
\keyword{internal}
