% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_x_axis_text_size}
\alias{get_x_axis_text_size}
\title{Internal helper function for getting a dynamic textsize for the ticks on the x axis. If there is a high amount of
ticks (either because a low step size or a big overall time range in the data) those texts can overlap if their size
is to big.}
\usage{
get_x_axis_text_size(breaks)
}
\arguments{
\item{breaks}{\verb{[numeric(1+)]} (if "By study day" is selected) or \verb{[Date(1+)]} (if "By study date" is selected) vector
setting the values at which ticks on this axis appear.}
}
\value{
\verb{[numeric(1)]} That is passed to the axis.text.x parameter of the ggplot2::theme() function.
}
\description{
Internal helper function for getting a dynamic textsize for the ticks on the x axis. If there is a high amount of
ticks (either because a low step size or a big overall time range in the data) those texts can overlap if their size
is to big.
}
\keyword{internal}
