% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{get_x_scale}
\alias{get_x_scale}
\title{Internal helper function for getting the breaks and labels for the x axis}
\usage{
get_x_scale(time_selection, x_axis)
}
\arguments{
\item{time_selection}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based) Either TIMETYPE_BUTTONS_OPT1 or TIMETYPE_BUTTONS_OPT2.}

\item{x_axis}{\verb{[list]} A list returned by get_x_axis(). Composed of:
\itemize{
\item breaks \verb{[numeric(1+)]}
A numeric vector setting the values at which ticks on this axis appear. Only contained in the list if the time
selection was "By study day".
\item ticktext \verb{[numeric(1+)]}
A character vector setting the text displayed at the ticks position via "breaks". Only contained in the list if the
time selection was "By study day".
\item breaks \verb{[Date(1+)]} Vector of the dates where the ticks should be in the plot. Only contained in the list if the
time selection was "By study date".
\item title \verb{[character(1)]} Not used in this function.
}}
}
\value{
a ggplot2 scale object used to configure the x-axis.
\itemize{
\item \verb{[ggplot2::ScaleContinuousPosition]}when time selection is "By study day" (numeric axis).
\item \verb{[ggplot2::ScaleDatePosition]} when time_selection is "By study date" (date axis).
This object can then be added to a ggplot to define tick positions and labels for the x-axis.
}
}
\description{
Internal helper function for getting the breaks and labels for the x axis
}
\keyword{internal}
