% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_validation.R
\name{is_valid_grouping_var}
\alias{is_valid_grouping_var}
\title{Internal helper function for checking if a variable is valid for grouping. The variable can't be null, has to
appear in the column names of the dataset passed to the function and must be of type factor to be considered as valid}
\usage{
is_valid_grouping_var(grouping_var, group_dataset)
}
\arguments{
\item{grouping_var}{\verb{[character(1) | NULL]} Character name of the variable that should be checked.}

\item{group_dataset}{\verb{[data.frame]} The data frame in which is looked for the grouping_var.}
}
\value{
logical value. TRUE if the grouping_var is valid, otherwise FALSE.
}
\description{
Internal helper function for checking if a variable is valid for grouping. The variable can't be null, has to
appear in the column names of the dataset passed to the function and must be of type factor to be considered as valid
}
\keyword{internal}
