% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_report_rates_app.R
\name{mock_report_rates_app}
\alias{mock_report_rates_app}
\title{Mock app for the Reporting Rates Module without the usage of the module manager.}
\usage{
mock_report_rates_app()
}
\description{
Launches an example app that shows a Reporting Rates module. The displayed
data is from the \pkg{pharmaversesdtm} package.
}
