% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_rr_app_with_mm.R
\name{mock_rr_app_with_mm}
\alias{mock_rr_app_with_mm}
\title{Mock app for the Reporting Rates Module integrated in the module manager.}
\usage{
mock_rr_app_with_mm()
}
\description{
Launches an example app that shows a Reporting Rates module integrated in the module manager surface. The displayed
data is from the \pkg{pharmaversesdtm} package.
}
\keyword{mock}
