% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_rates.R
\name{mod_report_rates}
\alias{mod_report_rates}
\title{Combining UI and server within DaVinci module}
\usage{
mod_report_rates(
  module_id,
  dm_dataset_name = "dm",
  ds_dataset_name = "ds",
  ae_dataset_name = "ae",
  subjid_var = REPORT_RATES$DEFAULTS$SUBJID,
  tooltip_decimal_places = REPORT_RATES$DEFAULTS$TT_DECIMAL_PLACES,
  disposition_events = list(event_var = REPORT_RATES$DEFAULTS$DISPO_VAR, date_var =
    REPORT_RATES$DEFAULTS$DISPO_DATE_VAR, day_var = REPORT_RATES$DEFAULTS$DISPO_DAY_VAR,
    entry_vals = REPORT_RATES$DEFAULTS$DISPO_ENTRY_VALS, exit_vals =
    REPORT_RATES$DEFAULTS$DISPO_EXIT_VALS),
  adverse_events = list(date_var = REPORT_RATES$DEFAULTS$AE_DATE_VAR, day_var =
    REPORT_RATES$DEFAULTS$AE_DAY_VAR),
  grouping_vars = list(choices = NULL, default_choice = NULL),
  x_step_size_list = list(days = REPORT_RATES$DEFAULTS$STEPSIZE_DAYS, weeks =
    REPORT_RATES$DEFAULTS$STEPSIZE_WEEKS, months = REPORT_RATES$DEFAULTS$STEPSIZE_MONTHS)
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique module ID to create a namespace.}

\item{dm_dataset_name}{\verb{[character(1)]}

Name of the Demographics dataset. Defaults to "dm".}

\item{ds_dataset_name}{\verb{[character(1)]}

Name of the Disposition dataset. Defaults to "ds".}

\item{ae_dataset_name}{\verb{[character(1)]}

Name of the Adverse Event dataset. Defaults to "ae".}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in all datasets
(default is USUBJID). Must be a single value.}

\item{tooltip_decimal_places}{\verb{[integer(1)]}

Number of decimal places to round the value shown in the hovertext. Defaults to 3L.}

\item{disposition_events}{\verb{[list]}

Named list specifying variables and values used to define entry and exit events in the disposition dataset.
Composed of:
\itemize{
\item event_var \verb{[character(1)]}
Name of the variable that contains the description of the event. Defaults to "DSDECOD".
\item date_var \verb{[character(1)]}
Name of the variable containing the date of the event. Defaults to "DSSTDTC".
\item day_var \verb{[character(1)]}
Name of the variable containing the study day of the event. Defaults to "DSSTDY".
\item entry_vals \verb{[character(1+)]}
Character vector of values defining which values out of var should be entry events. Defaults to c("RANDOMIZED").
\item exit_vals \verb{[character(1+)]}
Character vector of values defining which values out of var should be exit events.
Defaults to c("COMPLETED", "WITHDRAWAL BY SUBJECT", "DEATH").
}}

\item{adverse_events}{\verb{[list]}

Named list specifying the variables that define the date and study day of adverse event occurence.
Composed of:
\itemize{
\item date_var \verb{[character(1)]}
Name of the variable containing the date of the event. Defaults to "AESTDTC".
\item day_var \verb{[character(1)]}
Name of the variable containing the study day of the event. Defaults to "AESTDY".
}}

\item{grouping_vars}{\verb{[list]}

Named list specifying grouping options for the plot.
Composed of:
\itemize{
\item choices \verb{[character(1+) | NULL]}
Character vector of variable names out of the demographics dataset, specifying which variables are available for
grouping. If provided, only these variables can be selected. Defaults to \code{NULL}, meaning all factor variables
from the demographics dataset will be available for grouping.
\item default_choice \verb{[character(1) | NULL]}
Character specifying the default grouping variable to be used when starting the app. Defaults to \code{NULL}, meaning
no variable is selected for grouping by default.
}}

\item{x_step_size_list}{\verb{[list]}

Named list specifying the tick interval (step size) on the x-Axis depending on the time unit.
Composed of:
\itemize{
\item days \verb{[integer(1) | NULL]}
Integer value specifying the tick interval in days when the x-axis is displayed in days. If \code{NULL} it defaults
to a step size of 50L.
\item weeks \verb{[integer(1) | NULL]}
Integer value specifying the tick interval in weeks when the x-axis is displayed in weeks. If \code{NULL} it defaults
to a step size of 2L.
\item months \verb{[integer(1) | NULL]}
Integer value specifying the tick interval in months when the x-axis is displayed in months. If \code{NULL} it
defaults to a step size of 1L.
}}
}
\value{
A list containing the following elements to be used by the \pkg{modulemanager}:
\itemize{
\item \code{ui}: A UI function of the \pkg{dv.reportrate} module.
\item \code{server}: A server function of the \pkg{dv.reportrate} module.
\item \code{module_id}: A unique identifier.
\item \code{meta}: metadata.
}
}
\description{
This module shows plots of reporting rates of Adverse Events in clinical trials over time. There are 2 different
metrics that can be selected:
\itemize{
\item cumulative rate per total exposure time: Indicates the total number of events observed across all subjects,
divided by the sum of their individual observation times up to a given time point. This metric reflects the
overall event frequency relative to cumulative exposure.
\item reporting rates per active patients: Measures the number of reported events within a defined time interval
(e.g., week or month), relative to the number of patients actively under observation at the end of that interval.
The interval can be adjusted by selecting this rate type and specifying a bin size.
}
}
\keyword{main}
