% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_metric_calc.R
\name{prepare_ae_data}
\alias{prepare_ae_data}
\title{Internal helper function for preparing the ae dataframe by adding a column that indicates the occurrence of
adverse events.}
\usage{
prepare_ae_data(dataset, subjid_var, date_var, day_var)
}
\arguments{
\item{dataset}{\verb{[data.frame]} The Adverse Event dataframe.}

\item{subjid_var}{\verb{[character(1)]} Character name of the unique subject identifier column in all datasets.}

\item{date_var}{\verb{[character(1)]} Name of the variable containing the date of the event.}

\item{day_var}{\verb{[character(1)]} Name of the variable containing the study day of the event.}
}
\value{
A dataframe that contains an additional column "n" with the value 1 in every row, which indicates the
occurrence of an adverse event.
The column "m" with the only value 0 indicates that these rows don't describe Entry or Exit Events.
}
\description{
Internal helper function for preparing the ae dataframe by adding a column that indicates the occurrence of
adverse events.
}
\keyword{internal}
