% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_metric_calc.R
\name{prepare_ds_data}
\alias{prepare_ds_data}
\title{Internal helper function for preparing the ds dataframe by adding a column that indicates study entry or exit}
\usage{
prepare_ds_data(
  dataset,
  subjid_var,
  event_var,
  entry_terms,
  exit_terms,
  date_var,
  day_var
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} The disposition dataframe.}

\item{subjid_var}{\verb{[character(1)]} Character name of the unique subject identifier column in all datasets.}

\item{event_var}{\verb{[character(1)]} Name of the variable that contains the description of the event.}

\item{entry_terms}{\verb{[character(1+)]} Defining which values out of event_var should be entry events.}

\item{exit_terms}{\verb{[character(1+)]} Defining which values out of event_var should be exit events.}

\item{date_var}{\verb{[character(1)]} Name of the variable containing the date of the event.}

\item{day_var}{\verb{[character(1)]} Name of the variable containing the study day of the event.}
}
\value{
A dataframe that contains an additional column "m" which indicates study entry (m = 1) or exit (m = -1).
These markers are used in downstream calculations to derive the number of active subjects in the study at any time
point, and the cumulative exposure time.
The column "n" with the only value 0 indicates that these rows don't describe Adverse Events.
}
\description{
Internal helper function for preparing the ds dataframe by adding a column that indicates study entry or exit
}
\keyword{internal}
