% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Internal helper function for preparing the data for the plot depending on the selected inputs by the user (and the
rounding accuracy specified by the app creator). It orchestrates the sequence of helper functions that merge,
process and enrich the DS, AE and DM datasets into a unified, plot-ready dataframe with computed metrics, time
variables and hovertext.}
\usage{
prepare_plot_data(
  dm,
  ds,
  ae,
  subjid_var,
  grouping_var,
  time_selection,
  metric_selection,
  unit_selection,
  tt_dec_places
)
}
\arguments{
\item{dm}{\verb{[data.frame]} The demographics data frame.}

\item{ds}{\verb{[data.frame]} The disposition data frame returned by the prepare_ds_data() function.}

\item{ae}{\verb{[data.frame]} The adverse event data frame returned by the prepare_ae_data() function.}

\item{subjid_var}{\verb{[character(1)]} Character name of the unique subject identifier column in all datasets.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the dm data frame. The
values of this column define the different groups in the plot.}

\item{time_selection}{\verb{[character(1)]} Determining if the time on the x-axis should be study dates (calendar based)
or study days (subject based). Either TIMETYPE_BUTTONS_OPT1 or TIMETYPE_BUTTONS_OPT2.}

\item{metric_selection}{\verb{[character(1)]} For determining the calculated rate on the y-axis. Either
METRIC_BUTTONS_OPT1 or METRIC_BUTTONS_OPT2.}

\item{unit_selection}{\verb{[character(1)]} Determining the binsize when computing the reporting rate per active
patients. Either BINSIZE_BUTTONS_OPT1 or BINSIZE_BUTTONS_OPT2.}

\item{tt_dec_places}{\verb{[integer(1)]} Number of decimal places to round the value shown in the hovertext.}
}
\value{
A data frame containing processed data for plotting, including calculated metrics, time variables
and hover text.
}
\description{
Internal helper function for preparing the data for the plot depending on the selected inputs by the user (and the
rounding accuracy specified by the app creator). It orchestrates the sequence of helper functions that merge,
process and enrich the DS, AE and DM datasets into a unified, plot-ready dataframe with computed metrics, time
variables and hovertext.
}
\keyword{internal}
