% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hovertext.R
\name{remove_ht_from_dummy_rows}
\alias{remove_ht_from_dummy_rows}
\title{Internal helper function for removing the ht_time_base from the dummy rows that were only created to show the last
time interval correctly in the first place, and exchange it with "END" instead.}
\usage{
remove_ht_from_dummy_rows(dataset, grouping_var)
}
\arguments{
\item{dataset}{\verb{[data.frame]} Data frame that was passed to the get_hovertext_time function and processed by
the function.}

\item{grouping_var}{\verb{[character(1)]} Character name of the grouping variable (column) in the demographics data frame.
The values of this column define the different groups in the plot.}
}
\value{
A processed data frame without the ht_time_base and instead "END" in the dummy rows.
}
\description{
Internal helper function for removing the ht_time_base from the dummy rows that were only created to show the last
time interval correctly in the first place, and exchange it with "END" instead.
}
\keyword{internal}
