% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_rates.R
\name{report_rates_ui}
\alias{report_rates_ui}
\title{Create User interface for \pkg{dv.reportrates} module}
\usage{
report_rates_ui(module_id)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID provided to \code{report_rates_server()}}
}
\value{
A shiny UI.
}
\description{
Create User interface for \pkg{dv.reportrates} module
}
