% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{select_geom}
\alias{select_geom}
\title{Internal helper function for getting the right geom_..._interactive depending on the selected metric.}
\usage{
select_geom(metric_selection)
}
\arguments{
\item{metric_selection}{\verb{[character(1)]} For determining the calculated rate on the y-axis. Either
METRIC_BUTTONS_OPT1 or METRIC_BUTTONS_OPT2}
}
\value{
Either a ggiraph::geom_line_interactive or a ggiraph::geom_step_interactive function object.
}
\description{
Internal helper function for getting the right geom_..._interactive depending on the selected metric.
}
\keyword{internal}
