% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_validation.R
\name{validate_ae_dataset}
\alias{validate_ae_dataset}
\title{Internal helper function to perform checkmate validation checks on the adverse event dataset out of the
dataset_list() in a reactive context.}
\usage{
validate_ae_dataset(dataset, adverse_events)
}
\arguments{
\item{dataset}{\verb{[data.frame]} The adverse event data set}

\item{adverse_events}{\verb{[list]} Named list specifying the variables that define the date and study day of
adverse event occurrence. Composed of:
\itemize{
\item date_var \verb{[character(1)]}
Name of the variable containing the date of the event.
\item day_var \verb{[character(1)]}
Name of the variable containing the study day of the event.
}}
}
\description{
Internal helper function to perform checkmate validation checks on the adverse event dataset out of the
dataset_list() in a reactive context.
}
\keyword{internal}
