% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_validation.R
\name{validate_df_rows}
\alias{validate_df_rows}
\title{Internal helper function that checks if at least one of the data frames "dm", "ds" or "ae" contains zero rows OR
no levels are selected AND the "show ungrouped" checkbox isn't ticked while a grouping variable is selected.
If so, the UI shows a message "No data available", instead of the module proceeding with plot computation and showing
an empty coordinate system.}
\usage{
validate_df_rows(
  dataset_list,
  selected_levels,
  ungrouped_checkbox,
  selected_group
)
}
\arguments{
\item{dataset_list}{\verb{[list]} A list composed of:
\itemize{
\item dm \verb{[data.frame]} The dm dataframe.
\item ds \verb{[data.frame]} The ds dataframe.
\item ae \verb{[data.frame]} The ae dataframe.
}}

\item{selected_levels}{\verb{[character(1+) | NULL]} Character vector of the selected levels in the selected_group or NULL
if no levels are selected.}

\item{ungrouped_checkbox}{\verb{[logical]} Specifying if the ungrouped data should be displayed when a grouping variable
is selected.}

\item{selected_group}{\verb{[character(1)]} Character name of the selected grouping variable (column).}
}
\description{
Internal helper function that checks if at least one of the data frames "dm", "ds" or "ae" contains zero rows OR
no levels are selected AND the "show ungrouped" checkbox isn't ticked while a grouping variable is selected.
If so, the UI shows a message "No data available", instead of the module proceeding with plot computation and showing
an empty coordinate system.
}
\keyword{internal}
