% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_validation.R
\name{validate_dm_dataset}
\alias{validate_dm_dataset}
\title{Internal helper function to perform checks on the demographics dataset out of the dataset_list() and on the argument
that was passed to the grouping_vars parameter of the mod_report_rates() function, in a reactive context. Info
messages may be outputted in the console.}
\usage{
validate_dm_dataset(dataset, subjid_var, grouping_choices, default_choice)
}
\arguments{
\item{dataset}{\verb{[data.frame]} The demographics data set.}

\item{subjid_var}{\verb{[character(1)]} Character name of the unique subject identifier column in all datasets.}

\item{grouping_choices}{\verb{[character(1+) | NULL]} Character vector (of variable names out of dm), specifying which
variables are available for grouping.}

\item{default_choice}{\verb{[character(1) | NULL]} Character specifying the default grouping variable to be used. The
difference to grouping$default_choices is, that the argument that is passed to the default_choice param of this
function is a reactiveVal, which contains the default_choice in a reactive context (so either the value of
grouping$default_choice when first starting the app (given there was one provided), or NULL after that).
This ensures, that no unnecessary info messages will be outputted, because the default_choice is only used when first
starting the app.}
}
\description{
Internal helper function to perform checks on the demographics dataset out of the dataset_list() and on the argument
that was passed to the grouping_vars parameter of the mod_report_rates() function, in a reactive context. Info
messages may be outputted in the console.
}
\keyword{internal}
