% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_validation.R
\name{validate_grouping_selection}
\alias{validate_grouping_selection}
\title{Internal helper function for validating the current selected grouping variable.}
\usage{
validate_grouping_selection(dm, selected_group)
}
\arguments{
\item{dm}{\verb{[data.frame]} The demographics dataframe.}

\item{selected_group}{\verb{[character(1)]} Name of the current selected grouping variable}
}
\value{
Either the current selection of the grouping variable or REPORT_RATES$CHOICES$GROUP_NO_SELECTION if one of
the tests are FALSE.
}
\description{
Ensuring that when another demographics dataset gets selected, the grouping variable is reset to "-None-" if the
previous selected grouping variable isn't in the new dm dataset anymore or not of type factor anymore.
An informative message will we printed in the console if the grouping variable gets resetted.
}
\keyword{internal}
