% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{convert_to_girafe}
\alias{convert_to_girafe}
\title{Internal helper function to convert the ggplot object into an interactive girafe object. Applies all interactive
settings such as tooltip behaviour, hover effects, selection styling, zoom options, and toolbar configuration.}
\usage{
convert_to_girafe(ggplot_obj, selected_levels)
}
\arguments{
\item{ggplot_obj}{\verb{[ggplot2::ggplot]} The ggplot object generated by create_plot().}

\item{selected_levels}{\verb{[character(0+)]} Character vector containing the currently selected grouping levels that
should be displayed in the plot.}
}
\value{
A \verb{[ggiraph::girafe]} object. An interactive ggplot-based visualization showing calculated rates over time
based on the selected grouping variable and levels, metric, and time unit.
}
\description{
Internal helper function to convert the ggplot object into an interactive girafe object. Applies all interactive
settings such as tooltip behaviour, hover effects, selection styling, zoom options, and toolbar configuration.
}
