% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{generate_palette}
\alias{generate_palette}
\title{Internal helper function for generating a color palette, so that the color of the plots of the different levels stay
consistent when selecting or deselecting levels. For <= 9 levels (including the ungrouped data as an own level) the
"Okabe-Ito" palette is used. For <= 15 levels (including the ungrouped data as an own level) the "paletteMartin" is
used. Those palettes are colorblind-friendly. For > 15 levels  (including the ungrouped data as an own level) the
"viridis" palette is used.}
\usage{
generate_palette(all_levels)
}
\arguments{
\item{all_levels}{\verb{[character(0+)]} Vector of \emph{all} the available levels that are in the current grouping variable
column in the current demographics dataset. Which levels out of all available levels are selected in the UI dropdown
doesn't matter.}
}
\value{
A named character vector of color values, where the names correspond to the levels of the grouping variable.
}
\description{
Internal helper function for generating a color palette, so that the color of the plots of the different levels stay
consistent when selecting or deselecting levels. For <= 9 levels (including the ungrouped data as an own level) the
"Okabe-Ito" palette is used. For <= 15 levels (including the ungrouped data as an own level) the "paletteMartin" is
used. Those palettes are colorblind-friendly. For > 15 levels  (including the ungrouped data as an own level) the
"viridis" palette is used.
}
\keyword{internal}
