% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_rates.R
\name{report_rates_server}
\alias{report_rates_server}
\title{Create server for \pkg{dv.reportrate} module}
\usage{
report_rates_server(
  module_id,
  dataset_list,
  subjid_var = REPORT_RATES$DEFAULTS$SUBJID,
  tooltip_decimal_places = REPORT_RATES$DEFAULTS$TT_DECIMAL_PLACES,
  disposition_events,
  adverse_events,
  grouping_vars = list(choices = NULL, default_choice = NULL),
  x_step_size_list = list(days = REPORT_RATES$DEFAULTS$STEPSIZE_DAYS, weeks =
    REPORT_RATES$DEFAULTS$STEPSIZE_WEEKS, months = REPORT_RATES$DEFAULTS$STEPSIZE_MONTHS)
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID provided to \code{report_rates_ui()}}

\item{dataset_list}{\verb{[shiny::reactive(list(data.frame))]}

A reactive list of named datasets. Usually obtained from module manager.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in all datasets
(default is USUBJID). Must be a single value.}

\item{tooltip_decimal_places}{\verb{[integer(1)]}

Number of decimal places to round the value shown in the hovertext. Defaults to 3L.}

\item{disposition_events}{\verb{[list]}

Named list specifying variables and values used to define entry and exit events in the disposition dataset.
Composed of:
\itemize{
\item event_var \verb{[character(1)]}
Name of the variable that contains the description of the event. Defaults to "DSDECOD".
\item date_var \verb{[character(1)]}
Name of the variable containing the date of the event. Defaults to "DSSTDTC".
\item day_var \verb{[character(1)]}
Name of the variable containing the study day of the event. Defaults to "DSSTDY".
\item entry_vals \verb{[character(1+)]}
Character vector of values defining which values out of var should be entry events. Defaults to c("RANDOMIZED").
\item exit_vals \verb{[character(1+)]}
Character vector of values defining which values out of var should be exit events.
Defaults to c("COMPLETED", "WITHDRAWAL BY SUBJECT", "DEATH").
}}

\item{adverse_events}{\verb{[list]}

Named list specifying the variables that define the date and study day of adverse event occurence.
Composed of:
\itemize{
\item date_var \verb{[character(1)]}
Name of the variable containing the date of the event. Defaults to "AESTDTC".
\item day_var \verb{[character(1)]}
Name of the variable containing the study day of the event. Defaults to "AESTDY".
}}

\item{grouping_vars}{\verb{[list]}

Named list specifying grouping options for the plot.
Composed of:
\itemize{
\item choices \verb{[character(1+) | NULL]}
Character vector of variable names out of the demographics dataset, specifying which variables are available for
grouping. If provided, only these variables can be selected. Defaults to \code{NULL}, meaning all factor variables
from the demographics dataset will be available for grouping.
\item default_choice \verb{[character(1) | NULL]}
Character specifying the default grouping variable to be used when starting the app. Defaults to \code{NULL}, meaning
no variable is selected for grouping by default.
}}

\item{x_step_size_list}{\verb{[list]}

Named list specifying the tick interval (step size) on the x-Axis depending on the time unit.
Composed of:
\itemize{
\item days \verb{[integer(1) | NULL]}
Integer value specifying the tick interval in days when the x-axis is displayed in days. If \code{NULL} it defaults
to a step size of 50L.
\item weeks \verb{[integer(1) | NULL]}
Integer value specifying the tick interval in weeks when the x-axis is displayed in weeks. If \code{NULL} it defaults
to a step size of 2L.
\item months \verb{[integer(1) | NULL]}
Integer value specifying the tick interval in months when the x-axis is displayed in months. If \code{NULL} it
defaults to a step size of 1L.
}}
}
\description{
Create server for \pkg{dv.reportrate} module
}
